requirements.txt
appdirs==1.4.4
black==21.7b0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
google-api-core==1.31.1
google-auth==1.34.0
google-cloud-core==1.7.2
google-cloud-storage==1.42.0
google-crc32c==1.1.2
google-resumable-media==1.3.3
googleapis-common-protos==1.53.0
idna==3.2
mypy-extensions==0.4.3
numpy==1.21.1
oauthlib==3.1.1
packaging==21.0
pandas==1.3.1
pathspec==0.9.0
pkg_resources==0.0.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
tomli==1.2.1
urllib3==1.26.6
whetstone==0.3.1
